workspace/8_Kafka/requirements.txt
kafka-python
findspark